condition on OS & doesn't write to .bashrc

+++ ci/install-conda-env.sh
@@ -1,4 +1,6 @@
-#!/bin/bash -x
+#!/bin/bash
+
+set -ev
 
 halt() {
   echo $*
@@ -8,34 +10,38 @@ halt() {
 HERE="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 [ -z "$HOME" ] && export HOME="/root"
-sudo apt-get update
-sudo apt-get install -y  \
-    build-essential \
-    libatlas-base-dev \
-    liblapack-dev \
-    libffi-dev \
-    wget \
-    gfortran \
-    git \
-    libxft-dev \
-    libxpm-dev
+if [ `uname -s` != "Darwin "] ; then 
+    sudo apt-get update
+    sudo apt-get install -y  \
+        build-essential \
+        libatlas-base-dev \
+        liblapack-dev \
+        libffi-dev \
+        wget \
+        gfortran \
+        git \
+        libxft-dev \
+        libxpm-dev
+fi
 
-mkdir $HOME/.config/matplotlib -p
+mkdir -p $HOME/.config/matplotlib
 echo 'backend: agg' > $HOME/.config/matplotlib/matplotlibrc
 if [ -n "$TRAVIS_PYTHON_VERSION" ] ; then
     PENV_NAME="py${TRAVIS_PYTHON_VERSION:0:1}"
 else
-    PENV_NAME=`python --version 2>&1|awk '{print $2}'|awk -F . '{print $1}'`
-    PENV_NAME="py${PVERSION}"
+    PVERSION=`python --version 2>&1|awk '{print $2}'`
+    PENV_NAME="py${PVERSION:0:1}"
 fi
-wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O miniconda.sh
-chmod +x miniconda.sh
-export PATH=$HOME/miniconda/bin:$PATH
-./miniconda.sh -b && rm ./miniconda.sh || halt "Error installing miniconda"
+if ! which conda ; then
+    wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O miniconda.sh
+    chmod +x miniconda.sh
+    ./miniconda.sh -b && rm ./miniconda.sh || halt "Error installing miniconda"
+    export PATH=$HOME/miniconda/bin:$PATH
+fi
 conda env create --name $PENV_NAME --file $HERE/environment.yaml || halt "Error installing $PENV_NAME environment"
 source activate $PENV_NAME
 conda uninstall --yes gcc qt
-conda clean --yes -p -t
+conda clean --yes -p
 
 # install xgboost
 git clone https://github.com/dmlc/xgboost.git
@@ -54,8 +60,9 @@ python -c 'import ROOT, root_numpy' || halt "Error installing root_numpy"
 python -c 'import xgboost' || halt "Error installing XGboost"
 
 # environment
-cat << EOF > $HOME/.bashrc
+cat << EOF
+# add to your environment:
 export PATH=$HOME/miniconda/bin:$PATH
 source activate $PENV_NAME
-source '$ENV_BIN_DIR/thisroot.sh'
+pushd $ENV_BIN_DIR/.. ; source 'bin/thisroot.sh' ; popd
 EOF
